backend/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-enumchoicefield==2.0.0
django-environ==0.7.0
djangorestframework==3.12.4
psycopg2==2.9.1
PyJWT==2.3.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
